refactor(utils): share the timezone validator
