+++ packages/utils/src/index.ts
@@ -367,6 +367,7 @@ export {
   type UserTransitionTaskStatus,
   userTaskStatusTransitionRequiresComment,
 } from "./task-status-transitions.js";
+export { isValidTimezone } from "./timezone.js";
 export {
   extractBareHttpUrls,
   selectUnfurlCandidateUrls,
